// clients/js-kit/src/plugins/types.ts
export type BasePlugin = {
  authority: PluginAuthority;
  offset?: bigint;
};